requirements.txt
# LlamaIndex and core dependencies
llama-index==0.11.20
llama-index-embeddings-openai==0.2.5
llama-index-embeddings-gemini==0.4.1
llama-index-llms-openai==0.2.18
llama-index-llms-gemini==0.6.1
google-generativeai==0.8.0

# Vector databases
llama-index-vector-stores-milvus==0.1.21
pymilvus==2.3.6

# Caching
redis==5.0.1
# llama-index-storage-cache-redis==0.1.3  # Currently not available

# Document processing
llama-index-readers-file==0.1.31
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# Web framework (optional - for API interface)
fastapi==0.104.1
uvicorn==0.24.0

# Streamlit UI
streamlit==1.38.0
streamlit-chat==0.1.1
plotly==5.17.0
pandas==2.1.4

# Environment and utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.4.0
typing-extensions==4.9.0

# Additional utilities
nest-asyncio==1.5.8
aiofiles==23.2.1

# Streamlit UI
streamlit==1.38.0
plotly==5.17.0
pandas==2.1.4